requirements.txt
discord.py[voice]
python-dotenv
yt-dlp
ffmpeg-python
lyricsgenius
PyNaCl
requests
aiohttp